+++ src/Event/PayStart.php
@@ -4,10 +4,11 @@ declare(strict_types=1);
 
 namespace Yansongda\Pay\Event;
 
-use Yansongda\Pay\Contract\PluginInterface;
+use Yansongda\Artful\Contract\PluginInterface;
+use Yansongda\Artful\Event\Event;
 use Yansongda\Pay\Rocket;
 
-class PayStarted extends Event
+class PayStart extends Event
 {
     /**
      * @var PluginInterface[]