+++ tests/Direction/ResponseDirectionTest.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Yansongda\Pay\Tests\Direction;
-
-class ResponseDirectionTest extends NoHttpRequestDirectionTest
-{
-}